+++ java/src/main/java/ai/rapids/cudf/ColumnFilterOptions.java
@@ -31,26 +31,17 @@ public abstract class ColumnFilterOptions {
   // If empty all columns are returned.
   private final String[] includeColumnNames;
 
-  private final String[] decimal128Columns;
-
   protected ColumnFilterOptions(Builder<?> builder) {
     includeColumnNames = builder.includeColumnNames.toArray(
         new String[builder.includeColumnNames.size()]);
-    decimal128Columns = builder.decimal128Columns.toArray(
-        new String[builder.decimal128Columns.size()]);
   }
 
   String[] getIncludeColumnNames() {
     return includeColumnNames;
   }
 
-  String[] getDecimal128Columns() {
-    return decimal128Columns;
-  }
-
   public static class Builder<T extends Builder> {
     final List<String> includeColumnNames = new ArrayList<>();
-    final List<String> decimal128Columns = new ArrayList<>();
 
     /**
      * Include one or more specific columns.  Any column not included will not be read.
@@ -69,11 +60,6 @@ public abstract class ColumnFilterOptions {
      */
     public T includeColumn(Collection<String> names) {
       includeColumnNames.addAll(names);
-      return (T) this;
-    }
-
-    public T decimal128Column(String... names) {
-      decimal128Columns.addAll(Arrays.asList(names));
       return (T) this;
     }
   }
